django_app/requirements.txt
Django==4.2.7
psycopg2-binary==2.9.9
gunicorn==21.2.0